Play game bugs
Encountering issues when playing the game with a category with less than 5 questions. Because there is an assigned constant that is set to 5 to end the game when reached. Asking a mentor about an idea to assign a new variable to record the selected category # of questions to use as a game ender instead of 5. Incase there are not that many questions.